Update script.js
Updated displayHouse name to getHouse(). This function gets the houses without repeating any.
Added new displayHouse function. Still need to write the body of this function.
Added a 'houseList' variable and stored the getHouse() function there so I can access the function's values.
Added a displayDiv variable and set it equal to a div with the id of test. 
Wrote a for loop to display the houses / names as it loops over the houseList array.